Integrating add/remove licenses

+++ htm/controllers/update-qty.php
@@ -1,4 +1,5 @@
 <?php
+
 session_start();
 require 'config.php';
 require "../api/client/_init.php";
@@ -12,27 +13,33 @@ $subscriptionList = $subscription->getSubscriptionList();
 
 $sqlProvision = "SELECT is_provised from customer where id='$customer_id'";
 $resProvision = $conn->query($sqlProvision);
-if($resProvision->num_rows > 0){
+if ($resProvision->num_rows > 0) {
     while ($row = $resProvision->fetch_assoc()) {
         $provision = $row['is_provised'];
     }
-}
-else{
+} else {
     echo "Error: " . $sqlProvision . "<br>" . $conn->error;
 }
 echo "provision: " . $provision;
+echo "offerId: " . $subscriptionList->getOfferId();
 /*
-$sqlCartCheck = "SELECT * from cart where customer_id='$customer_id'";
-$resCartCheck = $conn->query($sqlCartCheck);
-if($resCartCheck->num_rows > 0){
-    $sqlDeleteCart = "DELETE from cart where customer_id='$customer_id'";
-    if($conn->query($sqlAddCompany) == True){
-        $sqlUpdateQty = "INSERT into cart set(customer_id, items, item_name, our_cost, msrp, proposed_cost, qty, transaction_id)"
-                . "VALUES($customer_id,)"
+if ($provision == 1) {
+    $subscriptionList[$i]->updateQuantity($qty);
+    header("Location:../portal/manageSubscription.php");
+} else {
+    $sqlCartCheck = "SELECT * from cart where customer_id='$customer_id'";
+    $resCartCheck = $conn->query($sqlCartCheck);
+    if ($resCartCheck->num_rows > 0) {
+        $sqlDeleteCart = "DELETE from cart where customer_id='$customer_id'";
+        if ($conn->query($sqlAddCompany) == True) {
+            $sqlUpdateQty = "INSERT into cart set(customer_id, items, item_name, our_cost, msrp, proposed_cost, qty, transaction_id)"
+            . "VALUES($customer_id, $subscriptionList[$i]->getOffer)";
+        }
     }
 }
-*/
 
-//$subscriptionList[$i]->updateQuantity($qty);
-//header("Location:../portal/manageSubscription.php");
+
+
+
+ */
 ?>
